refactor(tests): move fixture into their own files and organise better
