+++ packages/web-push/test/fixtures/vapid.ts
@@ -1,0 +1,7 @@
+// WARN: these keys are not secure, generate your own!
+export const insecureVapid = {
+  subject: 'mailto:test@test.test',
+  publicKey:
+    'BKKXE3jJV5UJ6c8HVPam6DvMPGZK26r-M7ojsO2T_KdjdeMT2d7oQpaO-VI3o3wn33mQ8JlHta3OSJ5f67Ac5ZY',
+  privateKey: 'VvFbnS5QlP5PcJzYyEpRQj93SFJetREtqqzbGAfK3RE',
+};
