+++ packages/web-push/test/tsconfig.json
@@ -5,7 +5,6 @@
     "noEmit": true,
     "composite": true
   },
-  "files": ["./vitest.config.ts"],
   "include": ["."],
   "references": [
     {
