+++ packages/web-push/test/vapid.test.ts
@@ -1,22 +1,10 @@
 import { afterEach, beforeEach, describe, test, vi } from 'vitest';
 import { vapidHeaders } from '../lib/vapid.js';
-import { fakeSubscriptions, fakeVapid } from './fixtures.js';
+import { subscriptions } from './fixtures/fixtures.js';
+import { insecureVapid } from './fixtures/vapid.js';
 
-describe('VAPID', () => {
-  beforeEach(() => {
-    // tell vitest we use mocked time
-    vi.useFakeTimers();
-  });
+test('Headers', async () => {
+  vi.setSystemTime(new Date(2000, 1, 1, 13));
 
-  afterEach(() => {
-    // restoring date after each test run
-    vi.useRealTimers();
-    vi.clearAllMocks();
-  });
-
-  test('Headers', async () => {
-    vi.setSystemTime(new Date(2000, 1, 1, 13));
-
-    await vapidHeaders(fakeSubscriptions.chrome, fakeVapid);
-  });
+  await vapidHeaders(subscriptions.chrome, insecureVapid);
 });
