+++ packages/web-push/vitest.config.ts
@@ -1,5 +1,7 @@
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  test: {},
+  test: {
+    setupFiles: ['./test/setup.ts'],
+  },
 });
